Add support for installing process-lib.
process-lib can now be installed via the install CMake target. Full
support is provided for finding process-lib with CMake's find_package
command after installing.

As part of this change the process-lib headers were also moved into
the process-lib directly (to be consistent when using an installed
version of process-lib). This was done to avoid conflicting with
other headers (process.h could already be in use by another project)