CMake: Install by default unless building object libraries.
Previously, we did not install static libraries when built from within a
subproject as the likely intention when building reproc as a static
library via `add_subdirectory` would be to not install reproc at all.
However, we now have `REPROC_OBJECT_LIBRARIES` which is a better
solution for this use case so we install static libraries by default
again.