reproc: Rework resource cleanup.
- Rename `reproc_handle` to `handle`. Possible because `reproc_handle`
is now private.
- Introduce `HANDLE_INVALID` (-1) which is a better invalid handle
compared to `NULL` and `0` as `0` is a valid file descriptor on POSIX
platforms.
- destroy functions now take the resource by value and return an invalid
instance of the resource instead of taking the resource by address and
setting it to an invalid resource value. The advantage of this approach
is that it's clearer at the call site that we're modifying the resource
since we assign the result of the destroy function to the resource.
- Fix a double free error when `reproc_start` fails.