reproc(++): Improve process exit and status handling.
- Remove `reproc_running`. Checking if a process is still running
involves system calls that can go wrong. It does not make sense to hide
these errors from users. `reproc_wait` with a timeout of zero can be
used to check if a process is still running.
- Make `reproc_wait` and `reproc_stop` return the child process exit
status if the child process has already exited or exits during the
course of these functions. This cuts out unnecessary boilerplate in the
examples and tests where we always immediately call `reproc_exit_status`
after `reproc_wait` and `reproc_stop`.
- Remove `reproc_exit_status`. `reproc_wait` with a timeout of 0 can be
used instead. This has the added advantage of checking if the process
has exited if it is still running.
- Check for `r == 0` in tests instead of `r >= 0`.
- When a process exits because of a signal, return the signal number
added to 256 as the process exit status.
- Introduce `REPROC_SIGTERM` and `REPROC_SIGKILL` to match against
signal exit statusses.
- Use `REPROC_SIGKILL` as the exit status for `TerminateProcess` on
Windows.
- Convert the Windows `CTRL-BREAK` exit status to `REPROC_SIGTERM`.
- Verify the correct signal exit status is returned in the stop test.
- Add corresponding constructs for all these changes to reproc++.
- Rename `exit_status` to `status` in reproc.c.
- Use separate constants for `status` in reproc.c instead of errors.
- Return a negative value for the return value when an error occurs and
`std::pair` or `std::tuple` is used as the return value.